pyproject.toml
[project]
name = "DroneObjDetectionGSU"
version = "0.1.0"
description = "Object Detection using Drone Imagery"

dependencies = ['opencv-python', 'numpy<2.0.0', 'ultralytics', 'tflite_runtime']

requires-python = ">=3.11"

[project.optional-dependencies]
pi = ['picamera2']
dev = ['mypy']
conversion = [
	'tensorflow-cpu',
	'tf_keras',
	'sng4onnx>=1.0.1',
	'onnx_graphsurgeon>=0.3.26',
	'onnx>=1.12.0',
	'onnx2tf>1.17.5,<=1.26.3',
	'onnxslim>=0.1.31',
	'tflite_support',
	'onnxruntime',
	'protobuf',
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.mypy]
ignore_missing_imports = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
warn_return_any = true
check_untyped_defs = true
cache_dir = ".cache/mypy"
